fix the semantic bug with toplevel function definition
We are currently hitting a number of lowering bugs with this approach,
so I have not yet been able to see how much code relies on this
undefined behavior. But my current theory is that very little code
should be relying on this, since it would often segfault.

Current bug is that `const x = y->y` tries to introduce an unexpected
local variable named `x` in addition to creating the global constant.
This was legal before since it was a no-op (though semantically
unsound), but now is will cause lowering to fail.

Secondly, we appear to be hitting an inference bug where it assumed
Expr(:toplevel) couldn't exist inside lowered code, even though that has
never been an actual restriction in the interpreter or codegen, but
which causes REPL completions to crash.